added background image, had my beautify run.
added backround.png and code for it in css. Comentted out previous js
code in the render function that drew the original background. My
beautify auto runs, so I let it, it removed all the trailing whitespace
and fixed some in dentation issues which can be seen in the changes as a
lot.